pyproject.toml
# SPDX-FileCopyrightText: 2025 CERN <home.cern>
#
# SPDX-License-Identifier: LGPL-2.1-or-later

[build-system]
requires = ["setuptools", "wheel", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "pts-framework"
dynamic = [
  "version",  # The version is taken directly from the repository, using setuptools-scm.
]
# description = "<Insert short description of the project>"
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
authors = [
  { name = "Odd Oyvind Andreassen", email = "odd.oyvind.andreassen@cern.ch" },
  { name = "Joseph Tagg", email = "joseph.tagg@cern.ch" },
  { name = "Bartlomiej Banas", email = "joseph.tagg@cern.ch" },
  { name = "Alvaro Martinez Landete", email = "alvaro.martinez.landete@cern.ch" },
]
# Specify the minimum Python version required. It is not safe to use
# this to control the *upper* version of Python.
requires-python = ">=3.11"

# The core dependencies of the project.
dependencies = [
  "hightime==0.2.2",
  "nidmm==1.4.8",
  "PyQt6",
  "PyYAML==6.0.2",
  # "numpy",
  # "other-mandatory-requirement",
]

[project.optional-dependencies]
# The optional dependencies of the project, installable using the "extras" syntax.
test = [
  "pytest",
]
doc = [
  "Sphinx",
  "sphinx",
  "acc-py-sphinx",
]
dev = [
  # Dependencies for development of the project, such as type checkers,
  # linters, etc.
  # "requirement-for-development-purposes-only",

  # Include the "test" and "doc" dependencies in the dev dependencies.
  "pts-framework[doc,test]",
  "ruff",
]

[project.urls]
# Homepage = "PyPTS repository"
Repository = "https://gitlab.cern.ch/pts/framework/pypts"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pypts", "pypts.*"]

[tool.setuptools.package-data]
# All package data which is tracked by git is included automatically due to
# include-package-data=true (the default) along with setuptools-scm. More details at:
# https://setuptools.pypa.io/en/latest/userguide/datafiles.html#include-package-data

[tool.setuptools_scm]
# Tell setuptools_scm to write a _version.py file in the package. The
# _version.py file must not be committed to the repository.
version_file = "src/pypts/_version.py"

[tool.acc_py_template]
features = [
  {name = "core", version = 2},
  {name = "docs", version = 2},
]
